# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iqrah Backend API",
    "description": "REST API for the Iqrah mobile app",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/google": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Google OAuth login handler.",
        "operationId": "google_auth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Returns current user profile from JWT context.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "API error response format",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "example": "[\"name: validation failed\"]"
          },
          "error": {
            "type": "string",
            "example": "Validation error: field is required"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "description": "Auth response with access token.",
        "required": [
          "access_token",
          "user_id",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "example": 3600,
            "minimum": 0
          },
          "user_id": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      },
...
      "GoogleAuthRequest": {
        "type": "object",
        "description": "Google OAuth login request.",
        "required": [
          "id_token"
        ],
        "properties": {
          "id_token": {
            "type": "string",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6Ii4uLiJ9..."
          }
        }
      },
...
      "UserId": {
        "type": "string",
        "description": "Stable user identifier.",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
      "UserProfile": {
        "type": "object",
        "description": "User profile response.",
        "required": [
          "id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T16:00:00Z"
          },
          "id": {
            "$ref": "#/components/schemas/UserId"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "example": "2026-02-20T16:10:00Z"
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}